Do not use saved window location if it is offscreen.
Problem statement: I ran into a problem when I copied a save from my desktop to my laptop where the desktop resolution was higher and it saved the window location offscreen.

Solution: Check to see if the window location is offscreen when the window is being opened. If it is offscreen, then just use the default location instead.